+++ python/sdss_access/__init__.py
@@ -15,6 +15,10 @@ from tree import Tree
 # File logging can be started by calling log.start_file_logger(name).
 from .misc import log
 
+# check if posix-based operating system
+from os import name
+is_posix = ( name == "posix" )
+
 # set up the TREE, but match the TREE_VER if it is already there
 from tree import Tree
 config = os.environ.get('TREE_VER', 'sdsswork')
@@ -22,7 +26,7 @@ tree = Tree(config=config)
 log.debug("SDSS_ACCESS> Using %r" % tree)
 
 from .path import Path as SDSSPath, AccessError
-from .sync import HttpAccess, RsyncAccess
+from .sync import HttpAccess, Access, BaseAccess, RsyncAccess, CurlAccess
 
 
 NAME = 'sdss_access'