+++ python/sdss_access/sync/cli.py
@@ -10,13 +10,18 @@ from tempfile import TemporaryFile
 from time import time, sleep
 from glob import iglob
 from datetime import datetime
+from sdss_access import is_posix
+from tempfile import gettempdir
 
 
 class Cli(object):
     """Class for providing command line interface (cli) sync scripts, and logs to local disk
     """
 
-    tmp_dir = '/tmp'
+    #tmp_dir = '/tmp'
+    tmp_dir = gettempdir()
+    if not is_posix:
+        tmp_dir = tmp_dir.split(':')[-1]
 
     def __init__(self, label=None, data_dir=None, verbose=False):
         self.label = label if label else 'sdss_access'
@@ -66,7 +71,7 @@ class Cli(object):
                 print("SDSS_ACCESS> [background]$ %r" % command)
             stdout = logfile if logfile else STDOUT
             stderr = errfile if errfile else STDOUT
-            background_process = Popen(split(str(command)), env=self.env, stdout=stdout, stderr=stderr)
+            background_process = Popen(split(str(command), posix=is_posix), env=self.env if 'rsync -' in command else None, stdout=stdout, stderr=stderr)
             if pause:
                 sleep(pause)
         else: